+++ engine/benchmarks/src/main/scala/pl/touk/nussknacker/engine/benchmarks/interpreter/InterpreterSetup.scala
@@ -52,7 +52,7 @@ class InterpreterSetup[T:ClassTag] {
     }
 
     val definitions = ProcessDefinitionExtractor.extractObjectWithMethods(configCreator,
-      api.process.ProcessObjectDependencies(ConfigFactory.empty(), ObjectNamingProvider(getClass.getClassLoader)))
+      api.process.ProcessObjectDependencies(ConfigFactory.empty(), ObjectNamingProvider(getClass.getClassLoader), RunMode.Engine))
 
     failOnErrors(CompiledProcess.compile(process, definitions, listeners, getClass.getClassLoader)(DefaultAsyncInterpretationValueDeterminer.DefaultValue))
   }