+++ engine/flink/process/src/main/scala/pl/touk/nussknacker/engine/process/compiler/VerificationFlinkProcessCompiler.scala
@@ -1,22 +1,22 @@
 package pl.touk.nussknacker.engine.process.compiler
 
-import org.apache.flink.api.common.ExecutionConfig
 import org.apache.flink.api.common.restartstrategy.RestartStrategies
 import org.apache.flink.streaming.api.scala._
 import pl.touk.nussknacker.engine.ModelConfigToLoad
 import pl.touk.nussknacker.engine.api.ProcessListener
 import pl.touk.nussknacker.engine.api.exception.EspExceptionInfo
 import pl.touk.nussknacker.engine.api.namespaces.ObjectNaming
-import pl.touk.nussknacker.engine.api.process.{ProcessConfigCreator, ProcessObjectDependencies}
+import pl.touk.nussknacker.engine.api.process.{ProcessConfigCreator, ProcessObjectDependencies, RunMode}
 import pl.touk.nussknacker.engine.definition.DefinitionExtractor
 import pl.touk.nussknacker.engine.flink.api.exception.FlinkEspExceptionHandler
 import pl.touk.nussknacker.engine.flink.util.source.EmptySource
 import pl.touk.nussknacker.engine.graph.EspProcess
 
-class VerificationFlinkProcessCompiler(process: EspProcess, executionConfig: ExecutionConfig,
-                                       creator: ProcessConfigCreator, config: ModelConfigToLoad,
+class VerificationFlinkProcessCompiler(process: EspProcess,
+                                       creator: ProcessConfigCreator,
+                                       config: ModelConfigToLoad,
                                        objectNaming: ObjectNaming)
-  extends StubbedFlinkProcessCompiler(process, creator, config, objectNaming) {
+  extends StubbedFlinkProcessCompiler(process, creator, config, objectNaming, RunMode.Verification) {
 
   override protected def listeners(processObjectDependencies: ProcessObjectDependencies): Seq[ProcessListener] = List()